reorganize the Dashboard and update axios
