+++ src/pages/Dashboard/Admin_Dashboard/UpdateUser/index.jsx
@@ -1,12 +1,12 @@
-import axios from "axios";
+import axiosInstance from "../../../../services/axios";
 import { useState } from "react";
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import Navbar from "../../../../components/Navbar/Navbar";
 
 export const UpdateUser = () => {
-  const baseURL = "http://localhost:3000";
-  const axiosInstance = axios.create({ baseURL });
+  // const baseURL = "http://localhost:3000";
+  // const axiosInstance = axios.create({ baseURL });
   const navigate = useNavigate();
   const [userData, updateUserData] = useState({
     userId: "",
