--- src/pages/Dashboard/Student_Dashboard/stu_dash_module.css
@@ -1,217 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Barlow&display=swap");
-
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  width: 200px;
-  background-color: #f1f1f1;
-  position: fixed;
-  height: 100%;
-  overflow: auto;
-}
-.header {
-  background-color: black;
-  padding: 20px;
-}
-
-li a {
-  display: block;
-  color: #000;
-  padding: 8px 0 8px 16px;
-  text-decoration: none;
-}
-li a:hover {
-  background-color: #f55;
-  color: white;
-}
-.active {
-  background-color: #4caf50;
-  color: white;
-}
-a {
-  text-decoration: none;
-}
-
-img {
-  float: left;
-}
-
-.dropbtn {
-  background-color: #4caf50;
-  color: white;
-  padding: 16px;
-  font-size: 16px;
-  border: none;
-  cursor: pointer;
-  padding: 10px 40px 10px 40px;
-}
-
-.dropdown {
-  position: relative;
-  display: inline-block;
-}
-
-.dropdown-content {
-  display: none;
-  position: absolute;
-  background-color: #f9f9f9;
-  min-width: 160px;
-  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-}
-
-.dropdown-content a {
-  color: black;
-  padding: 12px 16px;
-  text-decoration: none;
-  display: block;
-}
-
-.dropdown-content a:hover {
-  background-color: #f1f1f1;
-}
-
-.dropdown:hover .dropdown-content {
-  display: block;
-}
-
-.dropdown:hover .dropbtn {
-  background-color: #3e8e41;
-}
-#element1 {
-  float: left;
-}
-#element2 {
-  float: right;
-}
-.for-middle-text {
-  text-align: left;
-  float: left;
-  position: relative;
-  margin-left: 17%;
-}
-
-h2 {
-  font-family: "Franklin Gothic Medium", "Arial Narrow", Arial, sans-serif;
-  color: blue;
-}
-.my_pic {
-  width: 50px;
-  border-color: green;
-  border-width: 5px;
-  border-style: solid;
-  border-radius: 50%;
-}
-table th td {
-  border: 1px solid black;
-  border-collapse: collapse;
-}
-
-th td {
-  width: 30px;
-}
-.th_td {
-  padding: 5px;
-}
-td {
-  padding: 10px 25px 10px 25px;
-}
-
-/* search bar: */
-
-#tfheader {
-  background-color: #c3dfef;
-}
-#tfnewsearch {
-  float: right;
-  padding: 20px;
-}
-.tftextinput {
-  margin: 0;
-  padding: 5px 15px;
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 14px;
-  border: 1px solid #0076a3;
-  border-right: 0px;
-  border-top-left-radius: 5px 5px;
-  border-bottom-left-radius: 5px 5px;
-}
-.tfbutton {
-  margin: 0;
-  padding: 5px 15px;
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 14px;
-  outline: none;
-  cursor: pointer;
-  text-align: center;
-  text-decoration: none;
-  color: #ffffff;
-  border: solid 1px #0076a3;
-  border-right: 0px;
-  background: #0095cd;
-  background: -webkit-gradient(
-    linear,
-    left top,
-    left bottom,
-    from(#00adee),
-    to(#0078a5)
-  );
-  background: -moz-linear-gradient(top, #00adee, #0078a5);
-  border-top-right-radius: 5px 5px;
-  border-bottom-right-radius: 5px 5px;
-}
-.tfbutton:hover {
-  text-decoration: none;
-  background: #007ead;
-  background: -webkit-gradient(
-    linear,
-    left top,
-    left bottom,
-    from(#0095cc),
-    to(#00678e)
-  );
-  background: -moz-linear-gradient(top, #0095cc, #00678e);
-}
-/* Fixes submit button height problem in Firefox */
-.tfbutton::-moz-focus-inner {
-  border: 0;
-}
-.tfclear {
-  clear: both;
-}
-body {
-  font-family: "Barlow", sans-serif;
-}
-
-a:hover {
-  text-decoration: none;
-}
-
-.border-left {
-  border-left: 2px solid var(--primary) !important;
-}
-
-.sidebar {
-  top: 0;
-  left: 0;
-  z-index: 100;
-  overflow-y: auto;
-}
-
-.overlay {
-  background-color: rgb(0 0 0 / 45%);
-  z-index: 99;
-}
-
-/* sidebar for small screens */
-@media screen and (max-width: 767px) {
-  .sidebar {
-    max-width: 18rem;
-    transform: translateX(-100%);
-    transition: transform 0.4s ease-out;
-  }
-
-  .sidebar.active {
-    transform: translateX(0);
-  }
-}
