--- src/pages/Signin/signin.module.css
@@ -1,201 +1,0 @@
-@import url(https://fonts.googleapis.com/css?family=Varela);
-
-html {
-  height: 100%;
-}
-
-body {
-  background: #f2f2f2;
-  font-family: "Varela", sans-serif;
-  font-size: 14px;
-  line-height: 1.5;
-  color: #333;
-  min-height: 100%;
-  position: relative;
-}
-
-label {
-  margin-top: 6px;
-  line-height: 17px;
-}
-
-a {
-  color: #fff;
-}
-
-a:focus,
-a:hover {
-  color: #008080;
-}
-
-.checkbox-inline + .checkbox-inline,
-.radio-inline + .radio-inline {
-  margin-top: 6px;
-}
-
-/******* Login Page *******/
-
-body.login {
-  background: rgba(241, 242, 181, 1);
-  background: -moz-radial-gradient(
-    center,
-    ellipse cover,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(19, 80, 88, 1) 100%
-  );
-  background: -webkit-gradient(
-    radial,
-    center center,
-    0px,
-    center center,
-    100%,
-    color-stop(0%, rgba(255, 255, 255, 1)),
-    color-stop(100%, rgba(19, 80, 88, 1))
-  );
-  background: -webkit-radial-gradient(
-    center,
-    ellipse cover,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(19, 80, 88, 1) 100%
-  );
-  background: -o-radial-gradient(
-    center,
-    ellipse cover,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(19, 80, 88, 1) 100%
-  );
-  background: -ms-radial-gradient(
-    center,
-    ellipse cover,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(19, 80, 88, 1) 100%
-  );
-  background: radial-gradient(
-    ellipse at center,
-    rgba(255, 255, 255, 1) 0%,
-    rgba(19, 80, 88, 1) 100%
-  );
-  filter: progid: DXImageTransform.Microsoft.gradient(startColorstr='#f1f2b5', endColorstr='#135058', GradientType=1);
-}
-
-.relative {
-  position: relative;
-}
-
-.loginContainerWrapper .logo,
-.loginContainerWrapper .welcome {
-  margin: 0 0 20px 0;
-  font-size: 16px;
-  color: #fff;
-  text-align: center;
-  letter-spacing: 1px;
-}
-
-.loginContainerWrapper .logo {
-  text-align: center;
-  position: absolute;
-  top: -42px;
-  margin: 0 auto;
-  width: 25%;
-  left: 37.5%;
-  border-radius: 50%;
-  background-color: #344455;
-  padding: 25px;
-  box-shadow: 0px 0px 9px 2px #344454;
-}
-
-.loginContainerWrapper {
-  max-width: 400px;
-  margin: 10% auto 8%;
-  padding: 40px;
-  box-sizing: border-box;
-  background: rgba(57, 89, 116, 0.8);
-  box-shadow: 1px 1px 10px 1px #000000, 8px 8px 0px 0px #344454,
-    12px 12px 10px 0px #000000;
-  position: relative;
-  padding-top: 80px;
-}
-
-.logo .fa {
-  font-size: 50px;
-}
-
-.login input:focus + .fa {
-  color: #fff;
-}
-
-.loginForm .form-group {
-  margin-right: 0;
-  margin-left: 0;
-}
-
-.loginForm i {
-  position: absolute;
-  top: 18px;
-  right: 20px;
-  color: #93a5ab;
-}
-
-.loginForm .input-lg {
-  font-size: 16px;
-  height: 52px;
-  padding: 10px 25px;
-  border-radius: 0;
-}
-
-.login input[type="email"],
-.login input[type="password"],
-.login input:focus {
-  background-color: rgba(40, 52, 67, 0.75);
-  border: 1px solid #4a525f;
-  color: #eee;
-  border-left: 4px solid #93a5ab;
-}
-
-.login input:focus {
-  border-left: 4px solid #ccd8da;
-}
-
-input:-webkit-autofill,
-textarea:-webkit-autofill,
-select:-webkit-autofill {
-  background-color: rgba(40, 52, 67, 0.75) !important;
-  background-image: none;
-  color: rgb(0, 0, 0);
-  border-color: #faffbd;
-}
-
-.login .checkbox label,
-.login .checkbox a {
-  color: #ddd;
-}
-
-.btnSuccess {
-  background-color: transparent;
-  background-image: none;
-  padding: 8px 50px;
-  border-radius: 0;
-  border: 2px solid #93a5ab;
-  box-shadow: inset 0 0 0 0 #7692a7;
-  -webkit-transition: all ease 0.8s;
-  -moz-transition: all ease 0.8s;
-  transition: all ease 0.8s;
-}
-
-.btnSuccess:focus,
-.btnSuccess:hover,
-.btnSuccess.active,
-.btnSuccess:active {
-  background-color: transparent;
-  border-color: #fff;
-  box-shadow: inset 0 0 100px 0 #7692a7;
-  color: #fff;
-}
-
-#particlesJs {
-  /*   background: cornflowerblue; */
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  z-index: -1;
-}
