+++ src/services/axios.js
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const axiosInstance = axios.create({
+  baseURL: "http://localhost:3000",
+});
+
+export default axiosInstance;
